Move matchers to own package
